mfa: add cluster-level require_session_mfa option (#5939)
* mfa: add cluster-level require_session_mfa option

Users may choose to require MFA for all sessions in some clusters, or
only on select roles in others.
This PR adds the cluster-level option in the `auth_service` config.

To expose this field to non-auth services, opened up `GetAuthPreference`
access over the auth API, including all the cache plumbing.

* Address review feedback